+++ src/state/governor.rs
@@ -10,7 +10,7 @@ pub struct GovernorAccount {
     pda_data: PDAAccountData,
 
     /// The current fee-version (new requests are forced to use this version)
-    fee_version: u64,
+    fee_version: u32,
 
     /// The `ProgramFee` for the `FeeAccount` with the offset `fee_version`
     program_fee: ProgramFee,
@@ -18,7 +18,7 @@ pub struct GovernorAccount {
     /// The number of commitments in a MT-root hashing batch
     commitment_batching_rate: u32,
 
-    program_version: u64,
+    program_version: u32,
 }
 
 #[elusiv_account(pda_seed = b"sol_pool")]
